\name{list_poems}
\title{List available Emily Dickinson poems}
\usage{
list_poems()
}
\description{
Returns a vector of characters each of which represents the filename of a poem.  See https://github.com/Amherst-Statistics/DickinsonPoems for more information.
}
\value{
Returns a vector of character variables, one per poem.
}
\examples{
list_poems()
}
